+++ app/rts/.env.example
@@ -1,2 +1,3 @@
 APPSMITH_MONGODB_URI=mongodb://localhost:27017/appsmith
 APPSMITH_API_BASE_URL=http://localhost:8080/api/v1
+PORT=8091
