+++ app/rts/src/constants/socket.ts
@@ -1,0 +1,10 @@
+export const APP_ROOM_PREFIX = "app:";
+export const PAGE_ROOM_PREFIX = "page:";
+export const ROOT_NAMESPACE = "/";
+export const PAGE_EDIT_NAMESPACE = "/page/edit";
+
+export const EDITORS_EVENT_NAME = "collab:online_editors";
+export const START_EDIT_EVENT_NAME = "collab:start_edit";
+export const LEAVE_EDIT_EVENT_NAME = "collab:leave_edit";
+export const MOUSE_POINTER_EVENT_NAME = "collab:mouse_pointer";
+export const RELEASE_VERSION_EVENT_NAME = "info:release_version";
