+++ app/rts/src/constants/version.ts
@@ -1,0 +1,1 @@
+export const VERSION = "SNAPSHOT";
