--- app/rts/src/version.js
@@ -1,1 +1,0 @@
-export const VERSION = "SNAPSHOT";
